requirements.txt
altair==5.2.0
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
bing-image-urls==0.1.5
blinker==1.7.0
branca==0.7.1
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
cssselect==1.2.0
enum34==1.1.10
fake-useragent==1.4.0
folium==0.16.0
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
h2==4.1.0
haversine==2.8.1
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
logzero==1.7.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==2.21.0
mdurl==0.1.2
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parse==1.20.1
parsel==1.8.1
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.0
pydeck==0.8.1b0
pyee==11.1.0
Pygments==2.17.2
pyppeteer==2.0.0
pyquery==2.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.33.0
requests==2.27.1
requests-html==0.10.0
rfc3986==1.5.0
rich==13.7.1
rightmove-webscraper==1.1.2
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
streamlit==1.31.1
streamlit-folium==0.18.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
w3lib==2.1.2
websockets==10.4
xyzservices==2023.10.1
zipp==3.17.0
zoopla==1.0.0
geojson==3.1.0
plotly==5.20.0